// src/core/http/generated/models/whatsAppNotification.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Petvex Admin API
 * OpenAPI spec version: 1.0.0
 */
import type { Tenant } from './tenant.ts';
import type { WhatsAppNotificationAppointment } from './whatsAppNotificationAppointment.ts';
import type { WhatsAppNotificationClient } from './whatsAppNotificationClient.ts';
import type { WhatsAppNotificationPayload } from './whatsAppNotificationPayload.ts';
import type { WhatsAppNotificationPet } from './whatsAppNotificationPet.ts';
import type { WhatsAppNotificationStatus } from './whatsAppNotificationStatus.ts';

export interface WhatsAppNotification {
  id: string;
  tenant_id?: string;
  client_id?: string;
  appointment_id?: string;
  phone: string;
  template_key: string;
  status: WhatsAppNotificationStatus;
  error?: string;
  payload?: WhatsAppNotificationPayload;
  meta_message_id?: string;
  sent_at?: string;
  delivered_at?: string;
  read_at?: string;
  failed_at?: string;
  created_at: string;
  updated_at?: string;
  tenant?: Tenant;
  client?: WhatsAppNotificationClient;
  pet?: WhatsAppNotificationPet;
  appointment?: WhatsAppNotificationAppointment;
}